+++ src/pages/tenantPage.js
@@ -67,6 +67,7 @@ export default function TenantPage({
               <Grid item xs={12} sm={12} md={6} lg={4} xl={4}>
                 <DashboardCard
                   env={env}
+                  token={token}
                   index={index}
                   key={index}
                   pageType={